Multi_File_Type_BGQ_App
An improved version of the bgq insert Dash form. 
It enables the user to choose whether you want to insert a video file or an image file. Then, there's a selection  from various types of sources, and the app automatically scrapes the relevant meta data from each post.
In addition, it downloads the image and upload it to the relevant bucket.